feat: Add @tanstack/react-query dependency and update App component to use useQuery hook
This commit adds the "@tanstack/react-query" dependency to the project, with a version of "^5.62.11". It also updates the App component to use the useQuery hook from the "@tanstack/react-query" library. This hook is used to fetch photos and manage the loading state of the data. The fetchPhotos function is now wrapped in the useQuery hook, and the photoList state is replaced with the data and isLoading variables returned by the hook.